Create assets directory, move states and faq details in assets, update utils directory
