+++ src/assets/index.js
@@ -1,0 +1,2 @@
+export { default as StatesAndDistricts } from './statesAndDistricts';
+export { default as FAQ } from './faq';
