+++ src/components/About.js
@@ -1,48 +1,42 @@
 import Footer from './Footer';
 import { useState, useEffect } from 'react';
-import { FAQ } from '../utils/faq';
+import { FAQ } from '../assets';
 // TODO(slightlyoff): factor out common JSON parsing & caching of this file
 
 function About() {
-	const [faq, setFaq] = useState([]);
+  const [faq, setFaq] = useState([]);
 
-	useEffect(() => {
-		setFaq(FAQ.faq);
-	}, []);
+  useEffect(() => {
+    setFaq(FAQ.faq);
+  }, []);
 
-	useEffect(() => {
-		window.scrollTo(0, 0);
-	}, []);
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
 
-	// const getFAQs = () => {
-	// 	fetch(DATA_URL)
-	// 		.then((response) => {
-	// 			return response.json();
-	// 		})
-	// 		.then((data) => {
-	// 			setFaq(data.faq);
-	// 		})
-	// 		.catch((error) => {
-	// 			console.log(error);
-	// 		});
-	// };
+  return (
+    <>
+      <div className='About'>
+        {faq.map((faq, index) => {
+          return (
+            <div
+              key={index}
+              className='faq fadeInUp'
+              style={{ animationDelay: `${0.5 + index * 0.1}s` }}
+            >
+              <h2 className='question'>{faq.question}</h2>
+              <h2
+                className='answer'
+                dangerouslySetInnerHTML={{ __html: faq.answer }}
+              ></h2>
+            </div>
+          );
+        })}
+      </div>
 
-	return (
-		<>
-			<div className="About">
-				{faq.map((faq, index) => {
-					return (
-						<div key={index} className="faq fadeInUp" style={{ animationDelay: `${0.5 + index * 0.1}s` }}>
-							<h2 className="question">{faq.question}</h2>
-							<h2 className="answer" dangerouslySetInnerHTML={{ __html: faq.answer }}></h2>
-						</div>
-					);
-				})}
-			</div>
-
-			<Footer />
-		</>
-	);
+      <Footer />
+    </>
+  );
 }
 
 export default About;
