--- src/utils/faq.js
@@ -1,51 +1,0 @@
-export const FAQ = {
-	faq: [
-		{
-			answer: 'No.',
-			qno: '1',
-			question: 'Are you official?',
-		},
-		{
-			answer:
-				'We made every attempt to ensure the accuracy and reliability of the information provided. Kindly check and verify any/all information before acting upon the same. Always seek medical advice and refrain from self-diagnosis.',
-			qno: '1',
-			question: 'Disclaimer and Guide',
-		},
-		{
-			answer:
-				'We are aggregating data from social media, volunteer efforts and via the data collection on the website. We would love it if you can use this data in the fight against this virus. ',
-			qno: '2',
-			question: 'What are your sources? How is the data gathered for this project?',
-		},
-		{
-			answer:
-				'Once you add data via the form, one of our team members approves it before it comes on the website. This is done to mitigate against spam and dummy entries. If you want to add bulk data, kindly join our Discord server and we can get it done by the team.',
-			qno: '3',
-			question: 'How does the Add Data work?',
-		},
-		{
-			answer:
-				'The project is opensourced and welcoming contributors at <a href="https://github.com/getcovidhelpindia" target="_blank">github.com/getcovidhelpindia</a>',
-			qno: '4',
-			question: 'Where can I find the source code for this?',
-		},
-		{
-			answer:
-				'We started off as a group of friends NIT Rourkela who volunteered to do their part to help each other in this pandemic. We are a group of dedicated volunteers who curate and verify the data coming from several sources. ',
-			qno: '5',
-			question: 'Who are you?',
-		},
-		{
-			answer:
-				"Because it affects all of us. Today it's someone else who is getting infected; tomorrow it could be us. We need to support and aid everyone fighting this in all ways possible from our side.",
-			qno: '6',
-			question: 'Why are you guys putting in time and resources to do this while not gaining a single penny from it?',
-		},
-		{
-			answer:
-				'We need volunteers who can help us strengthen this initiative together. We would be delighted if you could add relavant data to this initiative via Add Info. We need help in terms of Data Collection, Data Verification, Front-End and Back-End Development, Social Media to name a few. We would be grateful if you join to help us. Join us at: <a href="http://discord.getcovidhelp.in/" target="_blank">discord.getcovidhelp.in/</a>',
-			qno: '7',
-			question: 'How can I help?',
-		},
-	],
-};
