+++ src/utils/loadScript.js
@@ -1,7 +1,9 @@
-export default loadScript = (src, done) => {
+const loadScript = (src, done) => {
   const js = document.createElement('script');
   js.src = src;
   js.onload = () => done();
   js.onerror = () => done(new Error('Failed to load the script: ', src));
   document.head.appendChild(js);
 };
+
+export default loadScript;
